fix(backup): move concurrency limiter into JobBackup*WithOptions
The semaphore was only in the scheduler wrappers, so GUI-triggered
backups bypassed the concurrency limit entirely. Move waitForBackupSlot
into JobBackupPhysicalWithOptions and JobBackupLogicalWithOptions so
all backup paths (scheduler, GUI, staging reseed) are gated.
